+++ src/main/java/org/sopt/service/PostValidator.java
@@ -1,0 +1,5 @@
+package org.sopt.service;
+
+public class PostValidator {
+
+}
